requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
whitenoise>=6.7,<7.0
Pillow>=11.0,<12.0
pytesseract>=0.3.10,<0.4.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
requests>=2.32,<3.0
deep-translator>=1.11,<2.0
